+++ tests/Jroc.Test262.Tests/built-ins/Function/prototype/bind/FunctionConformanceBatchExecutionTests.cs
@@ -1,0 +1,237 @@
+using Jroc.Test262.Tests.built_ins;
+
+namespace Jroc.Test262.Tests.built_ins.Function.prototype.bind;
+
+public class FunctionConformanceBatchExecutionTests : DiskExecutionTestsBase
+{
+    public FunctionConformanceBatchExecutionTests() : base("built_ins.Function.prototype.bind") { }
+
+    [Fact(DisplayName = "15.3.4.5-2-5.js")]
+    public Task _15_3_4_5_2_5() => ExecutionTestFromFile("15.3.4.5-2-5");
+
+    [Fact(DisplayName = "15.3.4.5-2-6.js")]
+    public Task _15_3_4_5_2_6() => ExecutionTestFromFile("15.3.4.5-2-6");
+
+    [Fact(DisplayName = "15.3.4.5-2-7.js")]
+    public Task _15_3_4_5_2_7() => ExecutionTestFromFile("15.3.4.5-2-7");
+
+    [Fact(DisplayName = "15.3.4.5-2-8.js")]
+    public Task _15_3_4_5_2_8() => ExecutionTestFromFile("15.3.4.5-2-8");
+
+    [Fact(DisplayName = "15.3.4.5-20-2.js")]
+    public Task _15_3_4_5_20_2() => ExecutionTestFromFile("15.3.4.5-20-2");
+
+    [Fact(DisplayName = "15.3.4.5-20-3.js")]
+    public Task _15_3_4_5_20_3() => ExecutionTestFromFile("15.3.4.5-20-3");
+
+    [Fact(DisplayName = "15.3.4.5-21-2.js")]
+    public Task _15_3_4_5_21_2() => ExecutionTestFromFile("15.3.4.5-21-2");
+
+    [Fact(DisplayName = "15.3.4.5-21-3.js")]
+    public Task _15_3_4_5_21_3() => ExecutionTestFromFile("15.3.4.5-21-3");
+
+    [Fact(DisplayName = "15.3.4.5-3-1.js")]
+    public Task _15_3_4_5_3_1() => ExecutionTestFromFile("15.3.4.5-3-1");
+
+    [Fact(DisplayName = "15.3.4.5-6-1.js")]
+    public Task _15_3_4_5_6_1() => ExecutionTestFromFile("15.3.4.5-6-1");
+
+    [Fact(DisplayName = "15.3.4.5-6-10.js")]
+    public Task _15_3_4_5_6_10() => ExecutionTestFromFile("15.3.4.5-6-10");
+
+    [Fact(DisplayName = "15.3.4.5-6-11.js")]
+    public Task _15_3_4_5_6_11() => ExecutionTestFromFile("15.3.4.5-6-11");
+
+    [Fact(DisplayName = "15.3.4.5-6-12.js")]
+    public Task _15_3_4_5_6_12() => ExecutionTestFromFile("15.3.4.5-6-12");
+
+    [Fact(DisplayName = "15.3.4.5-6-2.js")]
+    public Task _15_3_4_5_6_2() => ExecutionTestFromFile("15.3.4.5-6-2");
+
+    [Fact(DisplayName = "15.3.4.5-6-3.js")]
+    public Task _15_3_4_5_6_3() => ExecutionTestFromFile("15.3.4.5-6-3");
+
+    [Fact(DisplayName = "15.3.4.5-6-4.js")]
+    public Task _15_3_4_5_6_4() => ExecutionTestFromFile("15.3.4.5-6-4");
+
+    [Fact(DisplayName = "15.3.4.5-6-5.js")]
+    public Task _15_3_4_5_6_5() => ExecutionTestFromFile("15.3.4.5-6-5");
+
+    [Fact(DisplayName = "15.3.4.5-6-6.js")]
+    public Task _15_3_4_5_6_6() => ExecutionTestFromFile("15.3.4.5-6-6");
+
+    [Fact(DisplayName = "15.3.4.5-6-7.js")]
+    public Task _15_3_4_5_6_7() => ExecutionTestFromFile("15.3.4.5-6-7");
+
+    [Fact(DisplayName = "15.3.4.5-6-8.js")]
+    public Task _15_3_4_5_6_8() => ExecutionTestFromFile("15.3.4.5-6-8");
+
+    [Fact(DisplayName = "15.3.4.5-6-9.js")]
+    public Task _15_3_4_5_6_9() => ExecutionTestFromFile("15.3.4.5-6-9");
+
+    [Fact(DisplayName = "15.3.4.5-8-1.js")]
+    public Task _15_3_4_5_8_1() => ExecutionTestFromFile("15.3.4.5-8-1");
+
+    [Fact(DisplayName = "15.3.4.5-8-2.js")]
+    public Task _15_3_4_5_8_2() => ExecutionTestFromFile("15.3.4.5-8-2");
+
+    [Fact(DisplayName = "15.3.4.5-9-1.js")]
+    public Task _15_3_4_5_9_1() => ExecutionTestFromFile("15.3.4.5-9-1");
+
+    [Fact(DisplayName = "15.3.4.5-9-2.js")]
+    public Task _15_3_4_5_9_2() => ExecutionTestFromFile("15.3.4.5-9-2");
+
+    [Fact(DisplayName = "15.3.4.5.1-4-1.js")]
+    public Task _15_3_4_5_1_4_1() => ExecutionTestFromFile("15.3.4.5.1-4-1");
+
+    [Fact(DisplayName = "15.3.4.5.1-4-10.js")]
+    public Task _15_3_4_5_1_4_10() => ExecutionTestFromFile("15.3.4.5.1-4-10");
+
+    [Fact(DisplayName = "15.3.4.5.1-4-11.js")]
+    public Task _15_3_4_5_1_4_11() => ExecutionTestFromFile("15.3.4.5.1-4-11");
+
+    [Fact(DisplayName = "15.3.4.5.1-4-12.js")]
+    public Task _15_3_4_5_1_4_12() => ExecutionTestFromFile("15.3.4.5.1-4-12");
+
+    [Fact(DisplayName = "15.3.4.5.1-4-13.js")]
+    public Task _15_3_4_5_1_4_13() => ExecutionTestFromFile("15.3.4.5.1-4-13");
+
+    [Fact(DisplayName = "15.3.4.5.1-4-14.js")]
+    public Task _15_3_4_5_1_4_14() => ExecutionTestFromFile("15.3.4.5.1-4-14");
+
+    [Fact(DisplayName = "15.3.4.5.1-4-15.js")]
+    public Task _15_3_4_5_1_4_15() => ExecutionTestFromFile("15.3.4.5.1-4-15");
+
+    [Fact(DisplayName = "15.3.4.5.1-4-2.js")]
+    public Task _15_3_4_5_1_4_2() => ExecutionTestFromFile("15.3.4.5.1-4-2");
+
+    [Fact(DisplayName = "15.3.4.5.1-4-3.js")]
+    public Task _15_3_4_5_1_4_3() => ExecutionTestFromFile("15.3.4.5.1-4-3");
+
+    [Fact(DisplayName = "15.3.4.5.1-4-4.js")]
+    public Task _15_3_4_5_1_4_4() => ExecutionTestFromFile("15.3.4.5.1-4-4");
+
+    [Fact(DisplayName = "15.3.4.5.1-4-5.js")]
+    public Task _15_3_4_5_1_4_5() => ExecutionTestFromFile("15.3.4.5.1-4-5");
+
+    [Fact(DisplayName = "15.3.4.5.1-4-6.js")]
+    public Task _15_3_4_5_1_4_6() => ExecutionTestFromFile("15.3.4.5.1-4-6");
+
+    [Fact(DisplayName = "15.3.4.5.1-4-7.js")]
+    public Task _15_3_4_5_1_4_7() => ExecutionTestFromFile("15.3.4.5.1-4-7");
+
+    [Fact(DisplayName = "15.3.4.5.1-4-8.js")]
+    public Task _15_3_4_5_1_4_8() => ExecutionTestFromFile("15.3.4.5.1-4-8");
+
+    [Fact(DisplayName = "15.3.4.5.1-4-9.js")]
+    public Task _15_3_4_5_1_4_9() => ExecutionTestFromFile("15.3.4.5.1-4-9");
+
+    [Fact(DisplayName = "15.3.4.5.2-4-1.js")]
+    public Task _15_3_4_5_2_4_1() => ExecutionTestFromFile("15.3.4.5.2-4-1");
+
+    [Fact(DisplayName = "15.3.4.5.2-4-10.js")]
+    public Task _15_3_4_5_2_4_10() => ExecutionTestFromFile("15.3.4.5.2-4-10");
+
+    [Fact(DisplayName = "15.3.4.5.2-4-11.js")]
+    public Task _15_3_4_5_2_4_11() => ExecutionTestFromFile("15.3.4.5.2-4-11");
+
+    [Fact(DisplayName = "15.3.4.5.2-4-12.js")]
+    public Task _15_3_4_5_2_4_12() => ExecutionTestFromFile("15.3.4.5.2-4-12");
+
+    [Fact(DisplayName = "15.3.4.5.2-4-13.js")]
+    public Task _15_3_4_5_2_4_13() => ExecutionTestFromFile("15.3.4.5.2-4-13");
+
+    [Fact(DisplayName = "15.3.4.5.2-4-14.js")]
+    public Task _15_3_4_5_2_4_14() => ExecutionTestFromFile("15.3.4.5.2-4-14");
+
+    [Fact(DisplayName = "15.3.4.5.2-4-2.js")]
+    public Task _15_3_4_5_2_4_2() => ExecutionTestFromFile("15.3.4.5.2-4-2");
+
+    [Fact(DisplayName = "15.3.4.5.2-4-3.js")]
+    public Task _15_3_4_5_2_4_3() => ExecutionTestFromFile("15.3.4.5.2-4-3");
+
+    [Fact(DisplayName = "15.3.4.5.2-4-4.js")]
+    public Task _15_3_4_5_2_4_4() => ExecutionTestFromFile("15.3.4.5.2-4-4");
+
+    [Fact(DisplayName = "15.3.4.5.2-4-5.js")]
+    public Task _15_3_4_5_2_4_5() => ExecutionTestFromFile("15.3.4.5.2-4-5");
+
+    [Fact(DisplayName = "15.3.4.5.2-4-6.js")]
+    public Task _15_3_4_5_2_4_6() => ExecutionTestFromFile("15.3.4.5.2-4-6");
+
+    [Fact(DisplayName = "15.3.4.5.2-4-7.js")]
+    public Task _15_3_4_5_2_4_7() => ExecutionTestFromFile("15.3.4.5.2-4-7");
+
+    [Fact(DisplayName = "15.3.4.5.2-4-8.js")]
+    public Task _15_3_4_5_2_4_8() => ExecutionTestFromFile("15.3.4.5.2-4-8");
+
+    [Fact(DisplayName = "15.3.4.5.2-4-9.js")]
+    public Task _15_3_4_5_2_4_9() => ExecutionTestFromFile("15.3.4.5.2-4-9");
+
+    [Fact(DisplayName = "BoundFunction_restricted-properties.js")]
+    public Task BoundFunction_restricted_properties() => ExecutionTestFromFile("BoundFunction_restricted-properties");
+
+    [Fact(DisplayName = "S15.3.4.5_A1.js")]
+    public Task S15_3_4_5_A1() => ExecutionTestFromFile("S15.3.4.5_A1");
+
+    [Fact(DisplayName = "S15.3.4.5_A13.js")]
+    public Task S15_3_4_5_A13() => ExecutionTestFromFile("S15.3.4.5_A13");
+
+    [Fact(DisplayName = "S15.3.4.5_A14.js")]
+    public Task S15_3_4_5_A14() => ExecutionTestFromFile("S15.3.4.5_A14");
+
+    [Fact(DisplayName = "S15.3.4.5_A15.js")]
+    public Task S15_3_4_5_A15() => ExecutionTestFromFile("S15.3.4.5_A15");
+
+    [Fact(DisplayName = "S15.3.4.5_A16.js")]
+    public Task S15_3_4_5_A16() => ExecutionTestFromFile("S15.3.4.5_A16");
+
+    [Fact(DisplayName = "S15.3.4.5_A2.js")]
+    public Task S15_3_4_5_A2() => ExecutionTestFromFile("S15.3.4.5_A2");
+
+    [Fact(DisplayName = "S15.3.4.5_A3.js")]
+    public Task S15_3_4_5_A3() => ExecutionTestFromFile("S15.3.4.5_A3");
+
+    [Fact(DisplayName = "S15.3.4.5_A4.js")]
+    public Task S15_3_4_5_A4() => ExecutionTestFromFile("S15.3.4.5_A4");
+
+    [Fact(DisplayName = "instance-construct-newtarget-boundtarget-bound.js")]
+    public Task instance_construct_newtarget_boundtarget_bound() => ExecutionTestFromFile("instance-construct-newtarget-boundtarget-bound");
+
+    [Fact(DisplayName = "instance-construct-newtarget-boundtarget.js")]
+    public Task instance_construct_newtarget_boundtarget() => ExecutionTestFromFile("instance-construct-newtarget-boundtarget");
+
+    [Fact(DisplayName = "instance-construct-newtarget-self-new.js")]
+    public Task instance_construct_newtarget_self_new() => ExecutionTestFromFile("instance-construct-newtarget-self-new");
+
+    [Fact(DisplayName = "instance-construct-newtarget-self-reflect.js")]
+    public Task instance_construct_newtarget_self_reflect() => ExecutionTestFromFile("instance-construct-newtarget-self-reflect");
+
+    [Fact(DisplayName = "instance-length-exceeds-int32.js")]
+    public Task instance_length_exceeds_int32() => ExecutionTestFromFile("instance-length-exceeds-int32");
+
+    [Fact(DisplayName = "instance-length-prop-desc.js")]
+    public Task instance_length_prop_desc() => ExecutionTestFromFile("instance-length-prop-desc");
+
+    [Fact(DisplayName = "instance-length-remaining-args.js")]
+    public Task instance_length_remaining_args() => ExecutionTestFromFile("instance-length-remaining-args");
+
+    [Fact(DisplayName = "instance-name-chained.js")]
+    public Task instance_name_chained() => ExecutionTestFromFile("instance-name-chained");
+
+    [Fact(DisplayName = "instance-name-error.js")]
+    public Task instance_name_error() => ExecutionTestFromFile("instance-name-error");
+
+    [Fact(DisplayName = "instance-name-non-string.js")]
+    public Task instance_name_non_string() => ExecutionTestFromFile("instance-name-non-string");
+
+    [Fact(DisplayName = "instance-name.js")]
+    public Task instance_name() => ExecutionTestFromFile("instance-name");
+
+    [Fact(DisplayName = "length.js")]
+    public Task length() => ExecutionTestFromFile("length");
+
+    [Fact(DisplayName = "not-a-constructor.js")]
+    public Task not_a_constructor() => ExecutionTestFromFile("not-a-constructor");
+
+}