+++ tests/Jroc.Test262.Tests/built-ins/Function/prototype/caller/FunctionConformanceBatchExecutionTests.cs
@@ -1,0 +1,12 @@
+using Jroc.Test262.Tests.built_ins;
+
+namespace Jroc.Test262.Tests.built_ins.Function.prototype.caller;
+
+public class FunctionConformanceBatchExecutionTests : DiskExecutionTestsBase
+{
+    public FunctionConformanceBatchExecutionTests() : base("built_ins.Function.prototype.caller") { }
+
+    [Fact(DisplayName = "prop-desc.js")]
+    public Task prop_desc() => ExecutionTestFromFile("prop-desc");
+
+}